docs/requirements.txt
sphinx>=7.0,<9
myst-parser>=3.0,<5
rinohtype>=0.5,<1
sphinx-rtd-theme>=2.0,<3
